denote merge readiness in default graffiti
Replace `-$versionBlob` in default graffiti with `/🐼` to indicate
merge readiness according to social consensus.

Our current default graffiti format currently uses 32 bytes, which is
right at the capacity of what is possible. If we ever have a .10 patch
in Oct/Nov/Dec, we would be surprised by an overflow.
```
Nimbus/v##.##.#-######-stateofus
```