track `el_offline` status in VC
Updates the validator client to track the new `el_offline` field in the
`/eth/v1/node/syncing` response.
https://github.com/ethereum/beacon-APIs/pull/290

If `el_offline` is present, the BN will be treated same as if it was
optimistically synced. Note that a BN could report that it is in sync,
but may still have intermittent connection issues to the EL, making it
impossible to `getPayload` from them since Capella.